requirements.txt
        DJANFO
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0

        REACT
axios == 0.21.1,
bootstrap == 4.6.0,
react == 17.0.2,
react-bootstrap == 1.6.1,
react-dom == 17.0.2,
react-router-bootstrap == 0.25.0,
react-router-dom == 4.0.0



